Backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-multipart
passlib[bcrypt]
bcrypt==3.2.0
python-jose[cryptography]
pydantic
pydantic-settings
python-dotenv
email-validator
